fixes #5212. testNearCacheFasterThanGoingToTheCluster is moved to nightly because of its nature it needs to be on real network and not critical enough to be tested in each pr
